+++ src/hydro_plant.cpp
@@ -11,7 +11,7 @@ public:
 
     int get_emissionsOutput(Emissions& emissions) override;
 
-    int get_productionCost(float &cost, float powerRequest) override;
+    int get_productionCost(double &cost, double powerRequest) override;
 
 private:
     HydroPlant(const HydroPlant&) = delete;
@@ -40,7 +40,7 @@ HydroPlant::get_emissionsOutput(Emissions& emissions)
 
 
 int
-HydroPlant::get_productionCost(float &cost, float powerRequest)
+HydroPlant::get_productionCost(double &cost, double powerRequest)
 {
     cost = powerRequest * _runCost / 3600;
 }
