+++ src/solar_plant.cpp
@@ -12,7 +12,7 @@ public:
 
     int get_emissionsOutput(Emissions& emissions) override;
 
-    int get_productionCost(float &cost, float powerRequest) override;
+    int get_productionCost(double &cost, double powerRequest) override;
 
 private:
     UtilityManager* _um;
@@ -46,7 +46,7 @@ SolarPlant::get_emissionsOutput(Emissions& emissions)
 
 
 int
-SolarPlant::get_productionCost(float &cost, float powerRequest)
+SolarPlant::get_productionCost(double &cost, double powerRequest)
 {
     cost = powerRequest * _runCost / 3600;
 }
