+++ src/wind_plant.cpp
@@ -12,7 +12,7 @@ public:
 
     int get_emissionsOutput(Emissions& emissions) override;
 
-    int get_productionCost(float &cost, float powerRequest) override;
+    int get_productionCost(double &cost, double powerRequest) override;
 
 private:
     UtilityManager* _um;
@@ -46,7 +46,7 @@ WindPlant::get_emissionsOutput(Emissions& emissions)
 
 
 int
-WindPlant::get_productionCost(float &cost, float powerRequest)
+WindPlant::get_productionCost(double &cost, double powerRequest)
 {
     cost = powerRequest * _runCost / 3600;
 }
